Restructure explainers
